+++ features/lockscreen/impl/src/main/kotlin/io/element/android/features/lockscreen/impl/pin/model/PinDigit.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.element.android.features.lockscreen.impl.setup.model
+package io.element.android.features.lockscreen.impl.pin.model
 
 sealed interface PinDigit {
     data object Empty : PinDigit
