+++ features/lockscreen/impl/src/main/kotlin/io/element/android/features/lockscreen/impl/setup/SetupPinView.kt
@@ -18,21 +18,12 @@
 
 package io.element.android.features.lockscreen.impl.setup
 
-import androidx.compose.foundation.background
-import androidx.compose.foundation.border
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.consumeWindowInsets
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.imePadding
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.rememberScrollState
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.foundation.text.BasicTextField
-import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.foundation.verticalScroll
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Lock
@@ -41,13 +32,10 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.input.KeyboardType
-import androidx.compose.ui.text.input.TextFieldValue
 import androidx.compose.ui.tooling.preview.PreviewParameter
 import androidx.compose.ui.unit.dp
 import io.element.android.features.lockscreen.impl.R
-import io.element.android.features.lockscreen.impl.setup.model.PinDigit
-import io.element.android.features.lockscreen.impl.setup.model.PinEntry
+import io.element.android.features.lockscreen.impl.components.PinEntryTextField
 import io.element.android.features.lockscreen.impl.setup.validation.SetupPinFailure
 import io.element.android.libraries.designsystem.atomic.molecules.IconTitleSubtitleMolecule
 import io.element.android.libraries.designsystem.components.button.BackButton
@@ -55,10 +43,7 @@ import io.element.android.libraries.designsystem.components.dialogs.ErrorDialog
 import io.element.android.libraries.designsystem.preview.ElementPreview
 import io.element.android.libraries.designsystem.preview.PreviewsDayNight
 import io.element.android.libraries.designsystem.theme.components.Scaffold
-import io.element.android.libraries.designsystem.theme.components.Text
 import io.element.android.libraries.designsystem.theme.components.TopAppBar
-import io.element.android.libraries.designsystem.theme.pinDigitBg
-import io.element.android.libraries.theme.ElementTheme
 
 @Composable
 fun SetupPinView(
@@ -154,72 +139,6 @@ private fun SetupPinFailure.title(): String {
     return when (this) {
         SetupPinFailure.PinBlacklisted -> stringResource(id = R.string.screen_app_lock_setup_pin_blacklisted_dialog_title)
         SetupPinFailure.PinsDontMatch -> stringResource(id = R.string.screen_app_lock_setup_pin_mismatch_dialog_title)
-    }
-}
-
-@Composable
-private fun PinEntryTextField(
-    pinEntry: PinEntry,
-    onValueChange: (String) -> Unit,
-    modifier: Modifier = Modifier,
-) {
-    BasicTextField(
-        modifier = modifier,
-        value = TextFieldValue(pinEntry.toText()),
-        onValueChange = {
-            onValueChange(it.text)
-        },
-        keyboardOptions = KeyboardOptions(keyboardType = KeyboardType.Number),
-        decorationBox = {
-            PinEntryRow(pinEntry = pinEntry)
-        }
-    )
-}
-
-@Composable
-private fun PinEntryRow(
-    pinEntry: PinEntry,
-    modifier: Modifier = Modifier,
-) {
-    Row(
-        modifier = modifier,
-        horizontalArrangement = Arrangement.spacedBy(8.dp, alignment = Alignment.CenterHorizontally),
-        verticalAlignment = Alignment.CenterVertically,
-    ) {
-        for (digit in pinEntry.digits) {
-            PinDigitView(digit = digit)
-        }
-    }
-}
-
-@Composable
-private fun PinDigitView(
-    digit: PinDigit,
-    modifier: Modifier = Modifier,
-) {
-    val shape = RoundedCornerShape(8.dp)
-    val appearanceModifier = when (digit) {
-        PinDigit.Empty -> {
-            Modifier.border(1.dp, ElementTheme.colors.iconPrimary, shape)
-        }
-        is PinDigit.Filled -> {
-            Modifier.background(ElementTheme.colors.pinDigitBg, shape)
-        }
-    }
-    Box(
-        modifier = modifier
-            .size(48.dp)
-            .then(appearanceModifier),
-        contentAlignment = Alignment.Center,
-
-        ) {
-        if (digit is PinDigit.Filled) {
-            Text(
-                text = digit.toText(),
-                style = ElementTheme.typography.fontHeadingMdBold
-            )
-        }
-
     }
 }
